add Vagrantfile
Snap-core has been broken on Windows for a few weeks, and the watch preview
server has always had problems on Windows as well, so I am adding a VM to get
this to compile everywhere. Getting a recent GHC on Ubuntu was harder than it
should have been, but now `vagrant up` should produce a VM with all tools and
dependencies compiled and installed.