remove list of thoughts
Most notes that I wrote down were beyond the scope of this post, and most are
not worth a post of their own. The issue of performance is interesting though,
as well as the comparison with error codes (and other languages). This is just
beyond the scope of this post now, but perhaps there is enough material for
another post in the future.