rename Error to KelvinError
This should be less confusing with the `Error` trait and `io::Error`. The
target audience of this post is probably not familiar with either type, so
I think I should not add to the confusion by naming yet another type `Error`.
On the other hand, it is idiomatic to not prefix/suffix type names with the
project/namespace, but for this post I think this is okay.