Add base template
This is the initial commit of a new adventure: an attempt to re-write
the static site generator that builds my blog. Perhaps Hakyll is too
complicated for me. Perhaps I like to reinvent the wheel. Perhaps I
prefer moustache templates over Hakyll's dollars, not because moustache
templates are so awesome, but just because I find the curly brackets
more visually pleasing. Perhaps I don't like the fact that Hakyll pulls
in half of Hackage and takes half an hour to compile. Perhaps I want to
gain some more experience with Haskell. Or perhaps I am crazy.

I hope the new version will be shorter and simpler. At the very least
more readable. Of course that will not be the case. But I have to try.