Simplify context type
Context has to be a map, and the values can either be strings or other
pieces of context. It makes no sense to apply a template with a context
that is not a map, and it also makes no sense to put other things than
context in a list for my templating model.