Add basic front matter parser
This front matter parser is not robust at all. Haskell does not force
exhaustive pattern matching, so there is a lot that can go wrong. Still,
this strips off --- delimited front matter from a string, and parses it
into key value pairs. (Of type string. I do not need nor want YAML front
matter.)

It is not very hard to write a function that parses and strips the front
matter only if it is present (returning an empty map if there was no
front matter). But why bother? All my posts contain front matter, and if
I ever want to have a file without front matter it will be a special
case, so I can just omit the front matter extraction. Simpler is better.