Add function that expands dates
I will stop attaching times to my publication dates. Currently I do have
times on my publication dates (only machine-visible), but these times
are local times. For an international audience, local times do not make
sense. There are two possible resolutions:

 * Use local time plus time zone offset. I could retroactively figure
   out the correct offset for all my current posts and add those.
   Parsing dates plus time and offset is a bit more hairy than getting
   just the date.

 * Forget about time altogether, just set the date. (The exact time of
   publication would not be known.) This makes things simpler, but I
   also think it is the right thing to do. Often it will take a while
   after beginning to write a post until it is in a state where it can
   be published, and even then I often wait for a few days to fix typos
   and such. There is no well-defined time or even date at which a post
   is written. And from the point where I make the last change to a
   post, to the point where the html is online can take a few minutes,
   so having minute precision for the publication time makes no sense.

Considering this, posts will only have a publication date from now on,
not a publication time.

The function is a bit of a hack, in the sense that it has signature
`String -> String` and not `String -> Maybe String`. This is for my
personal use only, so I can get away with not handling errors properly.