Add function that adds computed metadata to posts
Computed metadata are things like alternative date formattings and the
url at which the post should end up.

Again, this function is a bit of a hack. There is no error handling for
when the "date" field is not present, but I can get away with that
because this is for my personal use only. Extracting the year is done by
dropping everything after the first hyphen -- not by properly parsing
the date and then formatting the year. It is not as robust, but it is
simpler.

Finally, the program is going in a "stringly-typed" direction at the
moment, where all post medadata is a string, and the field names are
strings too instead of actual fields of a Haskell data type. It is
quick and dirty. The disadvantage of swiching to something more
type-safe is that I would still need everything as strings to provide
the context for template expansion.