Refactor Post to be more type safe
The `Post` type is now a record type where the fields are set at parse
time. Other fields such as url and long-formatted date are now functions
instead of fields computed at parse time. In Haskell there is no
distinction between these functions or the field getters, which is nice.

The front matter is now used merely as an intermediate step to generate
the post. It will not be possible to set arbitrary fields in the front
matter and get them as context for template expansion. This means that
if I decide to modify a template and introduce an extra variable, I have
to go and change the Haskell code to propagate the field from front
matter to template expansion context.

The advantage of this is that the program will not be littered with map
lookups for hard-coded keys.