Add basic post template
Writing this, I think I need to change the conditionals in the
templating engine again. At the moment, my only use case is "include
this variable with something around it if it is set", and the other use
case that I will have ("include this if a variable is set") could be
supported in this way as well. The original reason for doing the string
comparison is no longer applicable now that there is an extra step
between front matter and expansion context.