Change base and post templates to be include-based
At first I did not like this idea, because it splits the main file into
two parts: the top (which contains the opening html tag) and bottom
(containing the closing tag). The "post" template would include both top
and bottom.

Instead, I wanted to have a single "base" template that contained the
top and bottom but not the content itself. Content would be just a
variable. Then there would be multiple steps of expansion, one to expand
a post into the base body, and one to expand base into the final result.
Unfortunately, when the parts are not nested, this gets very tedious. I
do not want to hard code the expansion steps for every type of page.

I see two options going forward:

 * Switch to an include-based system, and split the base template into
   top and bottom. There would be a single entry point for template
   expansion and the expansion path does not have to be hard coded. The
   "post" template can reuse the "head" template, and other pages can
   reuse that as well.

 * Keep using the nested templates, but replace their {{content}}
   variables with includes that automatically expand. This means the
   expansion path would not have to be hard-coded, but the "content"
   template would have to be different for every type of page.

For the first option, the template to expand differs per page type. For
the second option, the "content" variable differs per page type. Both
options can be implemented using the same mechanism in the templating
engine, so this is really a choice on how to write the templates. I
expect that the first option will lead to simpler code in the generator,
so that is what I will do for now.

The next step is to add support for includes to the templating engine,
and to make these includes available in the expansion context.