Add function that builds nested contexts
The templating engine does not really support having nested contexts in
the sense of a json object containing an object, but it is easy to fake
it by prepending something to the keys. For instance:

    {
      "foo":
      {
        "bar": "baz"
      }
    }

can be faked as

    {
      "foo.bar": "baz"
    }

The possibility of clashes is not a real concern, because the notation
"foo.bar" cannot reference both "bar" in "foo" and "foo.bar". Though
this is a hack, there is a certain elegance to it in my opinion. It adds
support for nested contexts in a very simple way.