Add function that picks related content for a post
Related content will appear below the post, in the footer. On my current
blog I have "previous post" and "next post" links, but I plan to replace
those with a "further reading" section that contains a peek of one other
post, inspired by the "now read this" feature of Svbtle and "don't stop
now" on Medium.

At the moment, the post that is selected is the next post chrono-
logically, or the previous post if there is no next post. I might add
manual hints to the front matter later, but this will do for now.

In addition to (or instead of, I am not sure yet) the further reading
section, I want to automatically group posts that are part of a series,
and provide links to all of the episodes. The type is there already, but
it is not used for now.