Add rudimental html minification function
The function works by parsing the html with the Tagsoup library, then
applying a strip function to the text inside the tags, and combining the
stripped pieces to a string again.

At the moment the stripping function is too optimistic. It should be
safe to strip whitespace after the opening tag or before the closing tag
of a block, but the current function does not take this into account. On
the other hand, perhaps the contribution of those spaces is so low that
it is not worth stripping them anyway.