Improve html minifier
This adds a lot of complexity, but I think this is unavoidable. The
previous minifier was just too aggressive. The current implementation
might still be too aggressive in some places, but I can fix issues as
I encounter them. For now, minification works surprisingly well. At
the moment process is as follows:

 1. Merge consecutive whitespace.
 2. Remove whitespace after opening tags.
 3. Remove whitespace before closing tags.

In contrast to the previous (overly simple) minifier, this does not
yet remove whitespace between non-inline tags. It also does not strip
comments.