Import posts from previous blog
This is almost a verbatim copy, but I dropped the date prefixes (more
on that below), I removed the times from the dates (see rationale in
previous commits), and I removed trailing whitespace and a discretionary
hyphen.

I want to avoid the redundancy of having the post date both in the
filename and in the front matter. Now that I am not using times any
more, all the information could be in the file name. So either it goes
in the filename, or it goes in the front matter.

 * In the filename:

   Pros: lexicographical order is chonological order.
   Cons: long filenames, harder to tab-complete, feels like magic.

 * In the front matter:

   Pros: all metadata in one place, more uniform parser.
   Cons: random order in file browsers.

I want to focus more on content and less on the time aspect of blogging.
Keeping the date in the front matter aligns better with that.