Add basic font subsetting script
There appears to be no font subsetting library for Haskell, so I will
have to include a small Python script to subset the webfonts. The
fonttools library is used for subsetting. The script is a small wrapper
that will be invoked from the Haskell executable.

PEP8 is unhappy about the spaces around named arguments, but readability
is more important in my opinion.