Add function to convert characters to glyph names
This will be used for font subsetting. Glyphs in a font are named with
postscript glyph names, so I can pass the glyphs to the subsetting
library. Actually the subsetting library accepts a piece of text too, in
addition to a list of glyph names, but I need to provide glyph names for
the ligatures anyway, so I might as well provide everything in terms of
glyph names.