Add @font-face in stylesheet for posts with code
If a post contains <code> tags, an @font-face rule will be included for
Inconsolata into a stylesheet in the head of the page.

In the previous version of my blog, @font-face rules were defined in a
stylesheet that was included by the html page. That means the browser
has to load the stylesheet before it can even begin loading the fonts.
By moving the rules to an inline stylesheet, loading can start
immediately after parsing the html.