Subset fonts to the fonts/ directory
A page id is assigned to every page and the filename of the subsetted
font is based on this page id. This way fonts can be referenced by an
absolute path, and it solves the problem of the root index, which would
not have a place to put the fonts with the old scheme. Furthermore, the
new font location does not depend on whether or not the final slash is
stripped from the url, which makes previewing locally easier.