Enable automatic expansions of punctuation
This is sometimes called "smart punctuation". It allows me to write "--"
in the markdown files, and it will turn into a real em-dash in the
rendered html.

The reason for doing this is not encoding. Everything is unicode and my
source files are UTF-8. The problem is that the different types of
dashes are very hard to distinguish with a monospace font. Even worse,
there is no way to tell whether a space is a thin space from an editor,
and thin spaces are a pain to insert manually. The idea is to make the
markdown as readable as possible when viewed unrendered in a monospace
font. Real quotes (“these” as opposed to the "straight" ones) work fine
with a monospace font, so I will enter those manually.

Current substitutions:

 * "foo -- bar" -> "foo — bar" with thin spaces around the em-dash.
 * "foo - bar" -> "foo – bar" with thin spaces around the en-dash.
 * "foo-bar" is left untouched.
 * "..." -> "…"

The ellipsis is pretty readable even in a monospace font. The reason for
doing the substitution here is pure laziness. On Windows one can insert
an ellipsis easily with the Colemak keyboard layout, but on Linux this
does not work out of the box, unfortunately.