Pull in css from redesign prototype
The current css is intended for small-screen devices, responsive scaling
to larger screens will be added in the future. The layout is based on a
baseline grid. At the moment only Firefox aligns elements properly to
this grid. Due to the way Blink and Webkit work, rounding errors
accumulate in those engines.