Add more postscript glyph names
The non-breaking space is a curious character. I do want a non-breaking
space in the post that uses it, but in an editor there is no way to see
that the space is non-breaking. Perhaps I should replace it with an
&nbsp; in the source. (And perhaps I should substitute that entity with
the unicode character, just like dashes and ellipses get transformed.)