Add support for body font
Enable determining the characters used in the body text, and subset the
body font. At the moment the subsetter is both not aggressive enough,
and it misses characters. Some more aggressive trimming will need to be
done to eliminate overlap with other fonts. Characters that are appended
via the css content property are not taken into account for subsetting.
These will have to be added manually later on.