Rewrite font handling
Instead of doing several passes over the content, and determining on
every pass which parts should be included for a specific font, do one
pass over the content and determine which font every part should have.
This is a bit simpler, and it ensures that no part is accidentally
included twice, for optimal subsetting. Furthermore, this makes
handling of small caps more flexible.