Refactor tag classification to use a stack
Instead of keeping nesting counts in a map, traverse all tags and keep
a stack of interesting ones. Apart from the classification, also keep
track of the tag name so elements can be popped when a tag is closed.
This will allow for classifications based on attributes in addition to
classification based on tag name.