Set 'read full post' link in serif small caps
At the moment this relies on the fact that the only link in the teaser
is the 'read full post' link, which is a bit of a hack. Currently, it is
not possible for a tag to be classified both by tag name and by
attribute. Perhaps a better way is to add a special class for serif
small caps.