Set run-in in small caps
The run-in incorrectly eats a space sometimes, and it does not look good
on every post. Perhaps I should enable them on a per-post basis, or
maybe I should not set run-ins in small caps after all. It might be a
bit over the top. On the other hand, for some posts it works really
well.