Add support for reponsive manual line breaks
Despite Adobe's efforts, balanced line breaking algorithms are still not
available in browsers. As a workaround, titles that require two lines
anyway can be broken manually.

If the viewport is wide enough, line breaks will be forced between the
words specified in the 'break' front matter. For narrow viewports this
has no effect.