Add script to losslessly reduce png size
The script applies various png compressors in succession. It trades
speed for compression ratio. Because I compress images only once, I can
afford to run the compressor for a few more minutes just to shave off a
few bytes. Any byte saved is a win.

This assumes the following tools are installed:

 * OptiPNG
 * PNGOut
 * Zopfli with ZofpliPNG