Copy compressed images to output directory
This does require the compressed images to be present in the first
place. This is bad, because it requires more than one build step. Either
I should put the compressed images under source control, and compress
every new image manually, or I should make jpeg compression part of the
generator.