Tweak font size in responsive design
This was quite a journey. Linear scaling with the viewport width seems
reasonable, but in practice it feels too large in many cases. I suspect
that this is because of the widescreen viewport: even as you try to fill
the entire screen but keep a moderate line length, only a small portion
of the page fits on the screen. This makes it hard to get an overview of
the page. It is better to leave some space at the sides.

I tried a lot of scaling functions, but in the end this square root
scaling works pretty well. I have yet to try it on larger screens
though.