Do not use math features of Pandoc
Instead of relying on Pandoc to read TeX math, I converted all of my
equations to raw html. This has several advantages over the MathML with
MathJax approach that I was using previously:

 * The resulting html file is smaller.

 * I have more control over the font that is used, and I can integrate
   more closely with my design.

 * There is no need to parse or typeset at the client. Everything
   renders instantly.

 * Equations can be wrapped by the browser if they do not fit on a line.

 * There is no need to download third-party js or webfonts.

And -- perhaps the biggest advantage of all -- this allows me to get rid
of all javascript on my blog.