Remove footer from index and contact pages
I tried once more (and succeeded) to get the sticky footer to work,
but it turns out that for Chrome on Android, the height of the browser
window is not constant. The address bar scrolls out of view normally
when you scroll down. If this happens on a page with height 100%, the
browser window grows higher, and the footer will jump.

All in all, it is not worth the trouble. The footer does not add any
value anyway. Let's get rid of it.