Import scale visualization
The design of my website follows a scale based on half-integer powers of
two. This works very well with Calluna, and as a mathematician it is a
nice detail.

The numbers have been rounded to one or two decimal places to minimize
rounding errors but still stick to the ideal scale to within a few
percent.