Fall back to normal caps if small caps are buggy
Blink had a bug before Chrome 48 that breaks small caps in combination
with letter spacing. I previously used all caps as a fallback, but that
is way too shouty. Instead, fall back to the original caps. The letter
spacing will still be applied, but that does not look as bad as the full
caps.