Add script to preprocess font files
The main purpose of this script is to add my custon designed glyphs, in
the fonts in fonts/extra/ to the orignal fonts. This is done with
FontForge, which has a very pleasant Python interface.

An unfortunate side effect of roundtripping a font through FontForge is
that it increases the file size. Even in subsetted fonts that should not
contain the extra glyph, the size of the woff2 files increases by
roughly half a kilobyte.

To counter this, I started looking into the tables that are altered by
the roundtrip, to see if their size can be reduced. The name table
is one of the tables affected, but after woff compression there is no
difference.