Strip type redundant cmap tables from fonts
One difference between the otf files before and after a FontForge
roundtrip, is that FontForge writes both a type 4 and a type 12 cmap
table. The type 12 table is redundant if all of the code points are in
the Basic Multilingual Plane (BMP). This is not the case for the double-
struck F glyph, but otherwise it is, so the extra cmap tables can be
removed.

Unfortunately, this is not the cause of the size increase of the woff
files, they are still about a kilobyte bigger than before the roundtrip.