Enable desubroutineization in subsetter
One of the major differences in the final woff2 files after subsetting,
between the one based on the original font, and the one that made the
FontForge roundtrip, was in CFF subroutines. Desubroutineization reduces
the file size greatly. In most cases, the file size decreased by a few
hundred bytes, in some cases even a kilobyte. In a few cases, the file
size increased, by a few hundred bytes at most.

This is a sufficient fix for the size problem introduced by the
FontForge roundtrip (and it also reduces font size for the fonts that
do not make the roundtrip), but it might be possible to shave off more
bytes. One thing that stood out is that FontForge inserts BlueShift
values that are not present in the original font.

Finally, the CFF table still contains the glyph names of all glyphs in
the font, even those excluded by subsetting. It should be possible to
remove those.

Before FontForge pipeline:

  woff mean:    6438 bytes
  woff median:  6728 bytes
  woff2 mean:   5575 bytes
  woff2 median: 6020 bytes

After this commit:

  woff mean:    6058 bytes
  woff median:  5886 bytes
  woff2 mean:   5192 bytes
  woff2 median: 5160 bytes