Fix subsetting of small caps
The css changes that removed the text-transform to uppercase broke small
caps, because if smcp is used instead of c2sc, the original lowercase
letters must be present instead of the uppercase ones. This fixes
subsetting to include the unaltered glyphs (so all the glyphs are there
for fallback if small caps are broken in Chrome pre-48), as well as
their small-cap variants.

A fortunate side-effect of this change is that the size of the webfonts
is reduced.