Upgrade Stackage resolver to LTS 7.1
This was supposed to be a small upgrade without behavioral changes, but
unfortunately highlighting-kate had severe breaking changes without
bumping the major version, and so most of the code snippets were broken
because the css class names had changed. Therefore the css for
highlighting is updated too in this commit. Because the names changed
only for some languages, the other definitions cannot be removed
unfortunately, leading to an increase in page size. More spans are
introduced by highlighting-kate too, leading to an extra page size hit.

Before:

  Type       Median Mean
  ---------- ------ ------
  html        11679  13112
  html gzip    4042   4319

  Mean page weight (woff):       42135
  Mean page weight (woff2):      37979
  Mean page weight (gz + woff):  33341
  Mean page weight (gz + woff2): 29185

After:

  Type       Median Mean
  ---------- ------ ------
  html        11691  13149
  html gzip    4049   4331

  Mean page weight (woff):       42171
  Mean page weight (woff2):      38016
  Mean page weight (gz + woff):  33353
  Mean page weight (gz + woff2): 29198

So the actual size hit is negligible. Good. Still, I should exclude the
css for code snippets on pages that don't have them, to make up for the
regression a bit. And I could do the same for tables.