Fix bug in archive classification
The classification was still relying on an id that was used previously,
but in the mean time this id had been replaced by a class. This caused
subsetting of the regular-weight serif font to be omitted.

Fortunately the only text in the serif font are the "read full post"
links, which do not change when the content of the page changes. I never
removed the files from my server, so they were still being served and
they had been subsetted correctly earlier. It was sheer luck that
nothing broke because of this though.

Many thanks to Amin Bandali for reporting this bug.