Upgrade Stackage resolver to LTS 8.20
This includes an update to Pandoc, which has a change in the way it
does syntax highlighting for Rust code. It now adds a <span class="op">
around things such as brackets in Rust code. It adds around 50 bytes to
the affected gzipped pages. It's a shame but not the end of the world.