Add postscript glyph names for mu and sigma
I need these for my post on benchmarking. Calluna does have a mu but not
a sigma, but I think I could draw it myself based on the o.

At the same time, the quote uses Calluna and not Calluna Sans, so change
variables to always be set in a sans-serif font, even inside a quote.