Replace favicon with one based on my avatar
It is not exactly the same as my avatar; the R has been snapped to pixel
boundaries for a 32x32 canvas, so it should render sharp even at small
sizes. Although the serifs align at a half pixel at 16x16, so they will
be a bit blurry at that size. Little can be done about that wihout
either violating alignment or distorting the shape too much.