Generate content names for subsetted fonts
Apart from being conceptually beautiful, this will solve caching issues
of updates. The downside is that generating html becomes a two-step
process: I have to know the content hash of the html to get the font
name, which is subsequently embedded in the css in the html head.

This commit introduces the content name computation, but it does not
pass those names into the css template.