Feed content-based font names into css template
The simplest way to do this at this point is a two-stage process: first
render the entire page without knowing the font filenames. From that we
can determine the glyphs that are used on the page. From the glyphs, the
font filenames can be computed as the hash of the glyphs used. This
filename is then fed into the template, and the page is rendered again.

This works because the font filenames do not appear in the visible part
of the page, so the glyphs will be the same in either case.

It makes the rendering process a bit slower, but that is an acceptable
price to pay.