Tweak runtime gc settings
As reported by ghc-gc-tune:

  4.78s:  +RTS -A8388608 -H268435456
  4.81s:  +RTS -A67108864 -H536870912
  4.82s:  +RTS -A33554432 -H536870912
  4.83s:  +RTS -A16777216 -H268435456
  4.83s:  +RTS -A536870912 -H8388608

So I've just written a rant about not including the variance of
measurements, and yet I am pasting these numbers here without anything
like that. It is indeed very likely that these settings are not really
optimal, as the timings are so close. But this is what ghc-gc-tune gives
me, so this is what I'll work with.