Fix bug font required check
The check would say that a font is used when there is a piece of text
on the page that uses the font. This applies to synthesized content as
well, but the way synthesized content was handled, was that it would
return an empty string if there is no content. And this empty string
caused the font to be marked as used. The solution is to filter empty
strings before checking whether a font is used.