Render posts in parallel
And -- also quite important -- compress the results in parallel too.
This cuts the render time for my entire blog almost in half. Runtime
options of -N4 still work best, better than -N2 or -N6 or -N8.

I might try having a work queue with as many compression processes as
cores, but for now this is a nice speedup that comes at very little cost
(in terms of lines of code).