Render svgs, so they can access font paths
This modifies the generator to extract image paths from the post
content, and copy the files, rather than a generic image copying step.
In the case of svg, it is rendered like a template with the post's font
context, so the svg can reference the body font. I need this for the
build time graph that has axis labels.

At the same time, get rid of the Config type in the generator and just
hard-code paths, it was a bit of premature over-engineering I think; I
never change the output path anyway.

There are two issues: first of all the subsetter is not aware of the
text in the svg so some glyphs are missing. And secondly, rendering the
template is very slow for the svg file. The latter I can probably
optimize, and for the former I'll add a hack to specify additional
glyphs.