Turn <img> tags into <object> tags for svg
Browsers do not allow embedded style sheets in svg in an <img> tags, but
they do allow this when it is inside an <object> tag. It seems I'll have
to go with an <object> tag then.

I will likely need to update my content security policy headers too to
allow this. Oh well, at least now I do get the right font in the graph.

I would expect that I would need to update my stylesheet as well to set
the right width, but so far everything just works without that.