Write conclusion about build environments
Writing things down is really a good exercise. Just writing this
paragraph made me realise very clearly the trade off between author and
user flexibility. One of the reasons that Blaze works so well for
Google, is because the author and consumer of the software are the same,
so pinning is not an issue. For software with many consumers the trade
offs are much more subtle.