Add css for <object> alignment
Tweaked such that the axis labels of my build systems graph align to the
baseline, the bars align to the baseline, and such that the content
after that graph is also properly baseline-aligned again.

This will trigger a change in all pages though, and also increase the
page size. I should conditionally include this based on whether there
are images or not, then it will at least be a reduction on most pages.