Add post about immutable data models
I started writing a Twitter post, but it did not fit a tweet. Then I
started doing a thread. I could have done it in 6 short tweets or so,
but at that point I realized I might as well turn it into a short blog
post. I haven't blogged in a while, and this is a nice little
observation that is worth sharing, and it only takes me 30 minutes or so
to write down, not days like I usually spend on my posts.

So here we go, finally a new post after more than a year of silence on
my blog.

Also, starting out writing in tweets was a good exercise to write
briefly and omit irrelevant details. I do like the extra breathing room
of a full post, but being forced to be brief helps to keep it to the
point.