Build the generator with Nix, not Stack
Stack has served me well for a long time, and it was a great tool that
fulfilled a pressing need when it was released. But nowadays Haskell
support in Nix is very good, and my generator binary is pretty simple.
I don't need Cabal or Stack, if Nix can provide me with a GHC with the
right packages installed.

A nice side effect is that the "blog" executable is now part of the
environment that you can enter with "nix run -c $SHELL", so there is no
second command needed to build it. And it is still possible to tamper
with it if I like.

Some modifications have to be made to facilitate this:

 * Nixpkgs contains a newer version of Pandoc than the one in the
   Stackage snapshot that I had pinned. This requires a few changes
   around configuring the reader and writer, and there's a new Pandoc
   monad that we have to use. The changes are fairly small though.

 * Locales and Nix is always problematic. But instead of relying on the
   current locale, the generator should be explicit about UTF-8, so do
   that instead, and then there are no locale problems either.

The GHC options from the Cabal file move into default.nix, and aside
from that, nothing much changes. After tweaking the Pandoc extensions,
the only difference in output now, as far as I can tell, are the line
number anchors included in code blocks. I suppose these are nice to
have to link to them, but they also increase the page size quite a bit,
and I'm not sure that is worth it, because I tend to have small
snippets, and my paragraph headings are linkable anyway. There does not
seem to be a way to turn them off at this point though, so I'll go with
it. A newer Pandoc version adds Ext_sourcepos, but I'm not sure if that
is the switch for it.