Begin writing post about shuffling
I wanted to add shuffle to Musium for a long time, I had thought about
how to do it (in a way that avoids consecutive plays of the same artist)
a few times without a clear conclusion. But today I went for a walk with
the goal of solving this, and after some consideration I think I do have
a neat solution, and a blog post would be a nice way to rubber duck that
it really works.

Of course Spotify already wrote this famous post, but after checking it
now, it is simpler than my algorithm, but it also does not prevent all
possible cases where an artist plays consecutively, which my algorithm
does. So it is still worth writing up!

The key idea: partition on artist, then interleave the two shortest
lists until you end up with one list.