Generalize to multiple artists
I still wonder, do we merge the smallest two partitions (including the
intermediate result), or merge the smallest into intermediate? After
thinking some more, the latter seems strictly better, it gives you more
“firepower” to split consecutive tracks apart, and I also cannot come up
with a counterexample.

Unrelated, for albums with multiple album artists. Do we count all
artists that are connected through an album as one partition (which
would avoid playing the same artist consecutively if possible, but which
may make it impossible due to having a larger partition)? Or do we treat
them as separate, at the risk of playing tracks that feature the same
artist consecutively? We could do both, first on narrow artist to build
the bigger partition, then we mix in the rest.