Allow some templating with m4
I end up writing a lot of <abbr> tags in my current post, and if I want
to be proper and not abuse <em> for variables, I should also use <var>
instead of <em>, which is annoying to write. m4 can help simplify this,
I think. I haven't previously used m4 but it looks simple enough, let's
give it a try.