Discover counterexample
This explains why I couldn’t prove my algorithm optimal ... it is not.
From the structure of the proof though, it is clear that it is possible
to build an optimal solution recurively. We have plenty of symbols to
achieve that, it’s just that it’s not obvious to me now how to do it in
a systematic and elegant way. An ugly way would be to first break up all
consecutive symbols, and we are guaranteed to succeed before we run out
of new symbols. Then insert the remainder arbitrarily. But that feels
ugly.