Write some observations and discussions
Now that I try this out with 500 tracks, I realize that it is a lot more
deterministic than it should be. Not such a good shuffle after all, or
rather. It can be rescued without modifying the definition of "optimal"
though, but that is because the algorithm outputs solutions that satisfy
a stronger definition (lower badness on *at least* one k, instead of on
*all* k).