Prevent Pandoc from changing the output
I want to compare the pre-GHC-update to the post-update, and Pandoc
changed the way it handles newlines, which caused every file to differ.
Fortunately we can turn this off.

Now the only difference is that some syntax highlighting .kw spans turn
into .cf spans, but I hightlighted them with the same color anyway, so
this is fine.