Update subsetter settings for new fonttools version
I was using a _really_ old version that needed many workarounts, the new
version resolves most of those, but without intervention also introduces
a regression in size. That was mostly caused by adding numr and dnom
glyphs even though they are unused, which were in turn caused by having
the layout closure. This can now be disabled, and doing so removes many
glyghs (for ligatures and small caps) that were not used, so the subset
fonts get quite a bit smaller on average, yay!