Allow more control over discretionary ligatures
I want to subtitle one post "The typechecker", and the ck ligature looks
pretty nice, but having the ch in there as well is overkill. So make a
span to turn them on at my ... discretion.

This does remove some discretionary ligatures where previously there was
an st ligature. It was a bit too much. I kept one.