Add a Ninja file to compress the svgs
Make is too clunky. I prefer writing Ninja files by hand over writing
Makefiles. Also, we should not compress actually because we need to
first render the svg, so the blog generator needs to do the compression.
But for now let's move this to Ninja.

Correction, the blog generator already renders and the compresses all
svg files, so there is nothing that I need to do in addition in
particular. No compression in Ninja then.